refactor(graduate): 스케줄 관련 API 및 훅을 새로운 구조로 통합 및 정리
