+++ apps/graduate/src/admin/entities/admin-schedule/api/index.ts
@@ -1,0 +1,2 @@
+export * from './fetchScheduleDetail';
+export * from './fetchScheduleList';
