+++ apps/graduate/src/admin/entities/admin-schedule/model/useScheduleDetail.ts
@@ -2,13 +2,13 @@ import { useQuery } from '@tanstack/react-query';
 
 import { KEYS } from '~/shared/constants';
 
-import { getScheduleDetail } from '~/admin/shared/api';
+import { fetchScheduleDetail } from '../api';
 
 export function useScheduleDetail(scheduleId: number) {
   return useQuery({
     queryKey: [...KEYS.SCHEDULE, scheduleId],
     queryFn: async () => {
-      const response = await getScheduleDetail(scheduleId);
+      const response = await fetchScheduleDetail(scheduleId);
       return response.data;
     },
     enabled: !!scheduleId,
