+++ apps/graduate/src/admin/entities/admin-schedule/model/useScheduleList.ts
@@ -2,13 +2,13 @@ import { useQuery } from '@tanstack/react-query';
 
 import { KEYS } from '~/shared/constants';
 
-import { getScheduleList } from '~/admin/shared/api';
+import { fetchScheduleList } from '../api/fetchScheduleList';
 
 export function useScheduleList() {
   return useQuery({
     queryKey: [...KEYS.SCHEDULE_LIST],
     queryFn: async () => {
-      const response = await getScheduleList();
+      const response = await fetchScheduleList();
       return response.data.contents ?? [];
     },
   });
