+++ apps/graduate/src/admin/pages/schedule/api/index.ts
@@ -1,0 +1,3 @@
+export * from './submitSchedule';
+export * from './updateSchedule';
+export * from './removeSchedule';
