+++ apps/graduate/src/admin/pages/schedule/api/removeSchedule.ts
@@ -1,0 +1,12 @@
+import { AxiosResponse } from 'axios';
+
+import { del } from '~/shared/api';
+import { END_POINT } from '~/shared/constants';
+
+export async function removeSchedule(
+  scheduleId: number,
+): Promise<AxiosResponse<void>> {
+  return del<void>({
+    request: END_POINT.ADMIN.SCHEDULE_DELETE(scheduleId),
+  });
+}
