+++ apps/graduate/src/admin/pages/schedule/model/useCreateSchedule.ts
@@ -3,13 +3,13 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { KEYS } from '~/shared/constants';
 import { CreateScheduleRequest } from '~/shared/types';
 
-import { createSchedule } from '~/admin/shared/api';
+import { submitSchedule } from '../api';
 
 export function useCreateSchedule() {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (data: CreateScheduleRequest) => createSchedule(data),
+    mutationFn: (data: CreateScheduleRequest) => submitSchedule(data),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [...KEYS.SCHEDULE_LIST] });
     },
