+++ apps/graduate/src/admin/pages/schedule/model/useDeleteSchedule.ts
@@ -2,13 +2,13 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { KEYS } from '~/shared/constants';
 
-import { deleteSchedule } from '~/admin/shared/api';
+import { removeSchedule } from '../api';
 
 export function useDeleteSchedule() {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (scheduleId: number) => deleteSchedule(scheduleId),
+    mutationFn: (scheduleId: number) => removeSchedule(scheduleId),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [...KEYS.SCHEDULE_LIST] });
     },
