+++ apps/graduate/src/admin/pages/schedule/model/useScheduleContent.ts
@@ -3,13 +3,13 @@ import { useQuery } from '@tanstack/react-query';
 import { KEYS } from '~/shared/constants';
 import type { SubmissionType } from '~/shared/types';
 
-import { getScheduleContent } from '~/admin/shared/api';
+import { fetchScheduleContent } from '../api/fetchScheduleContent';
 
 export function useScheduleContent(submissionType: SubmissionType) {
   return useQuery({
     queryKey: [...KEYS.SCHEDULE_CONTENT, submissionType],
     queryFn: async () => {
-      const response = await getScheduleContent(submissionType);
+      const response = await fetchScheduleContent(submissionType);
       return response.data;
     },
     enabled: !!submissionType,
