--- apps/graduate/src/admin/shared/api/index.ts
@@ -1,1 +1,0 @@
-export * from './schedule';
